+++ Tests/Tests/Collections/MemoryAllocator/V3/Dictionary.cs
@@ -4,14 +4,14 @@ namespace ME.ECS.Tests.MemoryAllocator.V3.Collections {
     using UnityEngine;
     using ME.ECS.Collections.MemoryAllocator;
 
-    public class HashSet {
+    public class Dictionary {
 
         [Test]
         public void Initialize() {
 
             var allocator = Base.GetAllocator(10);
 
-            var list = new HashSet<Vector3>(ref allocator, 10);
+            var list = new Dictionary<int, Vector3>(ref allocator, 10);
             Assert.IsTrue(list.isCreated);
             list.Dispose(ref allocator);
             
@@ -24,17 +24,17 @@ namespace ME.ECS.Tests.MemoryAllocator.V3.Collections {
 
             var allocator = Base.GetAllocator(10);
 
-            var list = new HashSet<Vector3>(ref allocator, 10);
+            var list = new Dictionary<int, Vector3Int>(ref allocator, 10);
             for (int i = 0; i < 100; ++i) {
 
-                list.Add(ref allocator, new Vector3(i, i, i));
+                list.Add(ref allocator, i, new Vector3Int(i, i, i));
                 
             }
 
             var cnt = 0;
             var e = list.GetEnumerator(in allocator);
             while (e.MoveNext() == true) {
-                Assert.IsTrue(e.Current.x >= 0 && e.Current.x < 100);
+                Assert.IsTrue(e.Current.Key == e.Current.Value.x && e.Current.Value.x >= 0 && e.Current.Value.x < 100);
                 ++cnt;
             }
             e.Dispose();
@@ -51,10 +51,10 @@ namespace ME.ECS.Tests.MemoryAllocator.V3.Collections {
 
             var allocator = Base.GetAllocator(10);
 
-            var list = new HashSet<Vector3>(ref allocator, 10);
+            var list = new Dictionary<int, Vector3>(ref allocator, 10);
             for (int i = 0; i < 100; ++i) {
 
-                list.Add(ref allocator, new Vector3(i, i, i));
+                list.Add(ref allocator, i, new Vector3(i, i, i));
                 
             }
             
@@ -70,16 +70,17 @@ namespace ME.ECS.Tests.MemoryAllocator.V3.Collections {
             var allocator = Base.GetAllocator(10);
 
             var cnt = 2;
-            var list = new HashSet<Vector3>(ref allocator, 10);
+            var list = new Dictionary<int, Vector3>(ref allocator, 10);
             for (int i = 0; i < cnt; ++i) {
 
-                list.Add(ref allocator, new Vector3(i, i, i));
+                list.Add(ref allocator, i, new Vector3(i, i, i));
                 
             }
             
             for (int i = 0; i < cnt; ++i) {
 
-                Assert.IsTrue(list.Contains(in allocator, new Vector3(i, i, i)));
+                Assert.IsTrue(list.ContainsKey(in allocator, i));
+                Assert.IsTrue(list.ContainsValue(in allocator, new Vector3(i, i, i)));
                 
             }
             
@@ -92,16 +93,16 @@ namespace ME.ECS.Tests.MemoryAllocator.V3.Collections {
 
             var allocator = Base.GetAllocator(10);
 
-            var list = new HashSet<Vector3>(ref allocator, 10);
+            var list = new Dictionary<int, Vector3>(ref allocator, 10);
             for (int i = 0; i < 100; ++i) {
 
-                list.Add(ref allocator, new Vector3(i, i, i));
+                list.Add(ref allocator, i, new Vector3(i, i, i));
                 
             }
             
             Assert.IsTrue(list.Count == 100);
-            Assert.IsTrue(list.Remove(ref allocator, new Vector3(50, 50, 50)));
-            Assert.IsFalse(list.Remove(ref allocator, new Vector3(50, 50, 50)));
+            Assert.IsTrue(list.Remove(ref allocator, 50));
+            Assert.IsFalse(list.Remove(ref allocator, 50));
             Assert.IsTrue(list.Count == 99);
 
             allocator.Dispose();
@@ -113,10 +114,10 @@ namespace ME.ECS.Tests.MemoryAllocator.V3.Collections {
 
             var allocator = Base.GetAllocator(10);
 
-            var list = new HashSet<Vector3>(ref allocator, 10);
+            var list = new Dictionary<int, Vector3>(ref allocator, 10);
             for (int i = 0; i < 100; ++i) {
 
-                list.Add(ref allocator, new Vector3(i, i, i));
+                list.Add(ref allocator, i, new Vector3(i, i, i));
                 
             }